Enhancements to Suggestion Feature and UI Updates (#23)
* Optimize suggestion filtering and UI updates

Refactored suggestion logic in `DocumentComponents` to delegate suggestion retrieval to a new model `StaticDocumentModel`, improving separation of concerns and encapsulation. This change includes filtering and mapping logic adjustments to cater to case-insensitive input matching and expanding the maximally shown suggestions to 100, enhancing the usability of the suggestion feature.

Removed obsolete `datalist` HTML elements and related DOM interactions for suggestion handling in `TextComponent` and `TextareaComponent`. Introduced a new `SuggestionComponent` class to centralize suggestion UI handling, allowing for a more structured and maintainable approach to rendering suggestions. This change improves the extensibility and encapsulation of suggestion rendering.

Commented out presentation and input container toggling in `BaseComponent` which suggests a move towards a different UI update strategy, possibly transitioning towards a pure CSS-based show/hide logic for elements during editing mode. Further investigation or follow-up changes would be needed to clarify and finalize this approach.

Restructured SCSS by moving editable data view styles to a dedicated partial, `_editableDataView.scss`, organizing the styles better and facilitating easier maintenance. This cleanup helps with the segregation of styles making the stylesheets more manageable.

It is still pending:
- Refactorization of the SuggestionComponent
- Conversion of the other components,
- and adjustments to the BaseComponent to be able to handle both types of components.

* Use fallback for missing document description

Modified document description retrieval to use a null-coalescing operator, ensuring an empty string is assigned when the description is unavailable. This change prevents potential exceptions during document component rendering when the description field is absent in the data model.

* Enhanced suggestion logic and UI toggle

Refactored the suggestion component to improve the clarity and maintainability of the code by introducing a structured comment system that details the functionality, parameters, and expected behavior of the component's methods and properties. This facilitation of understanding not only benefits current maintenance but also aids future developers in onboarding.

The key behavioral change was a shift in the suggestion index logic. Instead of relying on a boolean flag paired with a last index variable, the new implementation uses a dedicated index and a scrolling mode flag. The functional benefits of this are twofold: it simplifies the state management within the component and makes the suggestion cycling behavior clearer and more predictable.

On the UI side, edit mode interaction was polished, ensuring that presentation and input container toggling occurs efficiently upon mode transitions. The removal of commented-out code results in cleaner and more readable codebase.

To support these enhancements, unnecessary suggestion list manipulation methods were purged from the TextComponent, streamlining the component's responsibilities and leaning on the improved SuggestionComponent to handle suggestion-related tasks wholly.

All these changes collectively foster a more robust, streamlined, and understandable component architecture, optimizing the responsiveness of the suggestion system and the UI's edit/view mode transition, resulting in an improved UX.

* Version bump to V0.0.34