Feature/create-new-note-modal
* Add 'Create new note' translation keys

* Add Note Settings configuration options

Introduced UI elements for configuring default folder and template file for new notes in the settings tab, including input fields and related descriptions. Extended the settings interface to include a new noteSettings object, ensuring settings are saved and reloaded consistently. This addition enhances user customization for note creation within the plugin.

* Add getTagElements method to parse tag strings

* Add StaticNoteModel with filename generation

Introduced a new StaticNoteModel class providing a static API to generate filenames for notes. The filename creation process uses the note's date, tags, and title, formatting and sanitizing them as per application settings to ensure valid filenames that encapsulate note's metadata. This facilitates a standardized way of identifying notes and aids in organizing and retrieving note files.

* Add modal for creating new notes

Introduced a new feature with `CreateNewNoteModal` to allow users to create notes through a modal interface. The modal form consists of fields for the note's title, date, description, and tags. This class extends the `BaseModalForm` and includes methods for registering the command, opening the form, and evaluating form results to either create a new note or rename an existing one. The command registration is hooked into the main plugin lifecycle to ensure the command is available upon plugin initialization.

* Version bump to V0.0.47