Add toggle for preserving PDF file paths
Implemented a new option to prevent automatic changes to the PDF file paths during document processing. This includes the addition of a boolean field 'dontChangePdfPath' in the document interface and data model to enable users to opt-out of default PDF path adjustments. Enhanced the StaticDocumentModel to conditionally format and move PDF files based on this new setting. A toggle input was also added to the CreateNewMetadataModal to allow users to set this preference.