Implement deep cloning of FrontMatterCache for data integrity
Added a new utility method to deep clone FrontMatterCache objects, ensuring that the Obsidian Metadata Cache's data is not inadvertently mutated when consuming frontmatter data. The added function in the Helper class uses recursion to clone each property, handling both arrays and object literals. This update is applied within the FileModel to prevent direct modifications of the cached metadata, maintaining the integrity of the application's state.

Version bump to V0.0.53