Refactor acronym generation to model classes
Centralized responsibility for acronym generation from the CreateNewTaskManagementModal to the individual PrjTaskManagementModel and TaskModel classes. Removed redundant assignments and now directly use the getAcronym method after setting the title on the model. This change streamlines the modal's code by delegating the acronym creation to the specific model instances, which allows for overriding the default behavior in subclasses such as TaskModel, where a different acronym format is needed.