Add decorator for toString method control
Introduced a decorator function `toStringField` and the corresponding `ToStringField` Symbol, which allows developers to precisely mark which properties should be included when the `toString` method of a class is invoked. To facilitate this, an abstract base class `BaseData` has also been created that implements a common `toString` method. This enhancement provides fine-grained control over converters and formatters that rely on the `toString` method and avoids unnecessary or sensitive information exposure. Any class that extends `BaseData` will inherit this functionality and properties marked with `@toStringField` will be included in the `toString` output. This new approach allows for more modular and customizable output formatting.