Upgrade to tslib and typescript packages
This commit updates the versions of 'tslib' and 'typescript' packages as a part of our regular dependency update cycle. The 'tslib' version is bumped from a fixed 2.4.0 to a flexible ^2.6.2 version, and 'typescript' is upgraded from ^4.7.4 to ^5.4.5. These updates are aimed at overcoming any potential code issues that retrospectively added features/fixes in these packages may solve, ensuring better stability, and keeping up with the latest enhancements in TypeScript's evolving ecosystem.