Enhanced NoteData class and fixed CreateNewNoteModal handling
This commit tackles two major points.

First, it addresses an inconsistency when processing presets in the CreateNewNoteModal. A change has been made to skip entries where the preset value is a function.

Secondly, the NoteData class has been improved by extending it with the BaseData class and adding '@toStringField' decorators for class fields.