Refactor toString in models, add decorator to data classes
Replaced redundant string concatenation code in the NoteModel and PrjTaskManagementModel's toString methods with a single call to data's toString. This makes the toString method more maintainable and less prone to errors.

Incorporated a toStringField decorator in data classes (ProjectData, TaskData, TopicData) to mark fields to be included in the toString output. This allows fine-grained control over which fields are included in the string representation of data objects.

This change greatly improves readability and maintainability of our code, such changes are in line with our ongoing effort to clean up the project's codebase. It's not expected to have any effect on current functionality.